[Refactor] grid/feature-flag-types.h をsystem/enums/terrain/terrain-characterstics.h へ改名した
