+++ VisualStudio/Hengband/Hengband.vcxproj.filters
@@ -4265,9 +4265,6 @@
     <ClInclude Include="..\..\src\cmd-action\cmd-others.h">
       <Filter>cmd-action</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\grid\feature-flag-types.h">
-      <Filter>grid</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\src\action\run-execution.h">
       <Filter>action</Filter>
     </ClInclude>
@@ -5521,6 +5518,9 @@
       <Filter>system\floor</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\system\enums\terrain\terrain-tag.h">
+      <Filter>system\enums\terrain</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\system\enums\terrain\terrain-characteristics.h">
       <Filter>system\enums\terrain</Filter>
     </ClInclude>
   </ItemGroup>
