+++ src/combat/shoot.cpp
@@ -14,7 +14,6 @@
 #include "floor/geometry.h"
 #include "game-option/cheat-types.h"
 #include "game-option/special-options.h"
-#include "grid/feature-flag-types.h"
 #include "grid/feature.h"
 #include "grid/grid.h"
 #include "inventory/inventory-object.h"
@@ -49,6 +48,7 @@
 #include "system/artifact-type-definition.h"
 #include "system/baseitem/baseitem-key.h"
 #include "system/enums/monrace/monrace-id.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/item-entity.h"
