+++ src/dungeon/quest-completion-checker.cpp
@@ -4,12 +4,12 @@
 #include "floor/cave.h"
 #include "floor/floor-object.h"
 #include "floor/floor-util.h"
-#include "grid/feature-flag-types.h"
 #include "grid/feature.h"
 #include "grid/grid.h"
 #include "monster/monster-info.h"
 #include "object-enchant/item-apply-magic.h"
 #include "object-enchant/object-ego.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/item-entity.h"
