+++ src/effect/effect-processor.cpp
@@ -12,7 +12,6 @@
 #include "floor/line-of-sight.h"
 #include "game-option/map-screen-options.h"
 #include "game-option/special-options.h"
-#include "grid/feature-flag-types.h"
 #include "grid/grid.h"
 #include "io/cursor.h"
 #include "io/screen-util.h"
@@ -28,6 +27,7 @@
 #include "spell/range-calc.h"
 #include "system/angband-system.h"
 #include "system/enums/monrace/monrace-id.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/monrace/monrace-definition.h"
