+++ src/grid/feature-action-flags.h
@@ -1,6 +1,7 @@
 #pragma once
-#include "grid/feature-flag-types.h"
+
 #include "system/angband.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 
 /*
  * Terrain action flags
