+++ src/grid/feature-generator.cpp
@@ -8,11 +8,11 @@
 #include "game-option/cheat-types.h"
 #include "game-option/game-play-options.h"
 #include "grid/door.h"
-#include "grid/feature-flag-types.h"
 #include "room/lake-types.h"
 #include "room/rooms-builder.h"
 #include "system/dungeon/dungeon-data-definition.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/player-type-definition.h"
