+++ src/info-reader/feature-info-tokens-table.h
@@ -6,11 +6,10 @@
 
 #pragma once
 
-#include "grid/feature-flag-types.h"
-#include "system/angband.h"
 #include <string_view>
 #include <unordered_map>
 
+enum class TerrainCharacteristics;
 enum class TerrainTag;
 extern const std::unordered_map<std::string_view, TerrainCharacteristics> f_info_flags;
 extern const std::unordered_map<std::string_view, TerrainTag> terrain_tags;
