+++ src/mind/mind-elementalist.cpp
@@ -21,7 +21,6 @@
 #include "game-option/disturbance-options.h"
 #include "game-option/input-options.h"
 #include "game-option/text-display-options.h"
-#include "grid/feature-flag-types.h"
 #include "grid/grid.h"
 #include "hpmp/hp-mp-processor.h"
 #include "io/command-repeater.h"
@@ -55,6 +54,7 @@
 #include "status/bad-status-setter.h"
 #include "status/base-status.h"
 #include "system/enums/game-option-page.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/monrace/monrace-definition.h"
