+++ src/monster-floor/monster-lite.cpp
@@ -1,7 +1,6 @@
 #include "monster-floor/monster-lite.h"
 #include "dungeon/dungeon-flag-types.h"
 #include "floor/cave.h"
-#include "grid/feature-flag-types.h"
 #include "grid/grid.h"
 #include "monster-floor/monster-lite-util.h"
 #include "monster-race/race-brightness-flags.h"
@@ -11,6 +10,7 @@
 #include "player/special-defense-types.h"
 #include "system/angband-system.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/monrace/monrace-definition.h"
