+++ src/mspell/mspell-judgement.cpp
@@ -15,7 +15,6 @@
 #include "floor/cave.h"
 #include "floor/geometry.h"
 #include "floor/line-of-sight.h"
-#include "grid/feature-flag-types.h"
 #include "main/sound-definitions-table.h"
 #include "monster-race/race-flags-resistance.h"
 #include "monster/monster-status.h"
@@ -31,6 +30,7 @@
 #include "spell/range-calc.h"
 #include "system/angband-system.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/monrace/monrace-definition.h"
