+++ src/object-activation/activation-bolt-ball.cpp
@@ -4,10 +4,10 @@
 #include "effect/effect-processor.h"
 #include "floor/cave.h"
 #include "floor/floor-util.h"
-#include "grid/feature-flag-types.h"
 #include "hpmp/hp-mp-processor.h"
 #include "spell-kind/spells-launcher.h"
 #include "spell-kind/spells-specific-bolt.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/player-type-definition.h"
 #include "target/target-getter.h"
 #include "view/display-messages.h"
