+++ src/object-use/throw-execution.cpp
@@ -21,7 +21,6 @@
 #include "floor/floor-object.h"
 #include "floor/geometry.h"
 #include "game-option/cheat-types.h"
-#include "grid/feature-flag-types.h"
 #include "grid/grid.h"
 #include "inventory/inventory-object.h"
 #include "inventory/inventory-slot-types.h"
@@ -51,6 +50,7 @@
 #include "player/player-status-table.h"
 #include "racial/racial-android.h"
 #include "specific-object/torch.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/item-entity.h"
