+++ src/player-attack/player-attack.cpp
@@ -15,7 +15,6 @@
 #include "floor/cave.h"
 #include "floor/geometry.h"
 #include "game-option/cheat-types.h"
-#include "grid/feature-flag-types.h"
 #include "inventory/inventory-slot-types.h"
 #include "main/sound-definitions-table.h"
 #include "main/sound-of-music.h"
@@ -45,6 +44,7 @@
 #include "spell-kind/earthquake.h"
 #include "spell-realm/spells-hex.h"
 #include "sv-definition/sv-weapon-types.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/item-entity.h"
