+++ src/player-base/player-race.cpp
@@ -5,7 +5,6 @@
  * @details PlayerRaceからPlayerClassへの依存はあるが、逆は依存させないこと.
  */
 #include "player-base/player-race.h"
-#include "grid/feature-flag-types.h"
 #include "player-base/player-class.h"
 #include "player-info/mimic-info-table.h"
 #include "player/race-info-table.h"
