+++ src/realm/realm-hissatsu.cpp
@@ -13,7 +13,6 @@
 #include "effect/spells-effect-util.h"
 #include "floor/cave.h"
 #include "floor/geometry.h"
-#include "grid/feature-flag-types.h"
 #include "grid/feature.h"
 #include "grid/grid.h"
 #include "inventory/inventory-slot-types.h"
@@ -39,6 +38,7 @@
 #include "spell/technic-info-table.h"
 #include "status/bad-status-setter.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/item-entity.h"
