+++ src/room/treasure-deployment.cpp
@@ -7,12 +7,12 @@
 #include "room/treasure-deployment.h"
 #include "floor/cave.h"
 #include "floor/geometry.h"
-#include "grid/feature-flag-types.h"
 #include "grid/object-placer.h"
 #include "grid/trap.h"
 #include "monster-floor/monster-generator.h"
 #include "monster-floor/place-monster-types.h"
 #include "object-enchant/item-apply-magic.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/player-type-definition.h"
