+++ src/room/vault-builder.cpp
@@ -3,11 +3,11 @@
 #include "floor/floor-generator-util.h"
 #include "floor/floor-util.h"
 #include "game-option/cheat-options.h"
-#include "grid/feature-flag-types.h"
 #include "grid/object-placer.h"
 #include "grid/trap.h"
 #include "monster-floor/monster-generator.h"
 #include "monster-floor/place-monster-types.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/player-type-definition.h"
