+++ src/spell-kind/spells-fetcher.cpp
@@ -4,7 +4,6 @@
 #include "flavor/object-flavor-types.h"
 #include "floor/cave.h"
 #include "floor/geometry.h"
-#include "grid/feature-flag-types.h"
 #include "grid/grid.h"
 #include "monster-race/race-brightness-mask.h"
 #include "monster/monster-describer.h"
@@ -12,6 +11,7 @@
 #include "monster/monster-update.h"
 #include "monster/monster-util.h"
 #include "system/angband-system.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/item-entity.h"
