+++ src/spell-kind/spells-neighbor.cpp
@@ -4,9 +4,9 @@
 #include "effect/effect-processor.h"
 #include "floor/cave.h"
 #include "floor/floor-util.h"
-#include "grid/feature-flag-types.h"
 #include "grid/grid.h"
 #include "spell-kind/earthquake.h"
+#include "system/enums/terrain/terrain-characteristics.h"
 #include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
 #include "util/bit-flags-calculator.h"
