+++ src/system/enums/terrain/terrain-characteristics.h
